+++ tests/integration_tests/test_spatial_correlation.py
@@ -34,7 +34,7 @@ from probeye.inference.scipy_.correlation_models import \
 class TestProblem(unittest.TestCase):
 
     def test_spatial_correlation(self, n_steps=200, n_initial_steps=100,
-                                 n_walkers=20, plot=False, verbose=False,
+                                 n_walkers=20, plot=False, show_progress=False,
                                  run_scipy=True, run_emcee=True,
                                  run_torch=False):
         """
@@ -53,8 +53,8 @@ class TestProblem(unittest.TestCase):
             If True, the data and the posterior distributions are plotted. This
             is deactivated by default, so that the test does not stop until the
             generated plots are closed.
-        verbose : bool, optional
-            If True, additional information will be printed to the console.
+        show_progress : bool, optional
+            If True, progress-bars will be shown, if available.
         run_scipy : bool, optional
             If True, the problem is solved with scipy (maximum likelihood est).
             Otherwise, no maximum likelihood estimate is derived.
@@ -220,7 +220,8 @@ class TestProblem(unittest.TestCase):
                        'l_corr': l_corr}
         run_inference_engines(problem, true_values=true_values, n_steps=n_steps,
                               n_initial_steps=n_initial_steps,
-                              n_walkers=n_walkers, plot=plot, verbose=verbose,
+                              n_walkers=n_walkers, plot=plot,
+                              show_progress=show_progress,
                               run_scipy=run_scipy, run_emcee=run_emcee,
                               run_torch=run_torch)